Fix swiftlint violations
