+++ EduVPN-redesign/Controllers/iOS/ItemSelectionViewController.swift
@@ -90,4 +90,3 @@ extension ItemSelectionViewController {
         delegate?.itemSelectionViewController(self, didSelectIndex: indexPath.row)
     }
 }
-
